Fix signed/unsigned comp

+++ src/transform.cpp
@@ -198,7 +198,7 @@ namespace OpenBabel
           //number of atoms in the pattern (passed as a string in the option text)
           //has to be the same as in the molecule.
           itr2 = pOptions->find("exactmatch");
-          if(itr2!=pOptions->end() && NumHvyAtoms()!=atoi(itr2->second.c_str()))
+          if(itr2!=pOptions->end() && (int)NumHvyAtoms()!=atoi(itr2->second.c_str()))
             fmatch=false;
           else
             {
